moved quit data
